refactor(core): perform strictly-enforced structural DDD alignments
- Relocated spec/domains to core/scivis to respect core as the domain boundary.
- Eradicated 'toolkit' flat directory and correctly housed split-brain instances (flow_diff.py to core/workflow, gatekeeper to core/security, state_rewind to core/state, and visualizer to core/presentation) to abide by physical isolation of bounded contexts.
- Updated all import definitions throughout ports, core, and test suites to resolve paths to the correctly mapped directories.
- Mirrored tests suite folder structure to accurately reflect src structure.
- Removed unused imports and fixed minor formatting anomalies detected by ruff & mypy checks during the refactor.
