+++ src/coreason_manifest/core/presentation/visualizer.py
