+++ src/coreason_manifest/core/workflow/evals.py
@@ -81,5 +81,3 @@ class EvalsManifest(CoreasonModel, frozen=True, extra="forbid"):
 
 class TestCase(SimulationScenario):
     """Alias for SimulationScenario used in standard test runs."""
-
-    pass
