+++ tests/coreason_manifest/test_basic.py
@@ -7,4 +7,3 @@ def test_schema_import() -> None:
     import coreason_manifest
 
     assert coreason_manifest is not None
-
